requirements.txt
Flask>=3.0
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
Werkzeug>=3.0
requests>=2.31.0
python-docx>=1.1.0
reportlab>=4.1.0
pytest>=8.0.0
pytest-flask>=1.3.0
Flask-Session>=0.5.0
Flask-Cors>=4.0.0
